+++ packages/solana-wallet-snap/src/core/services/assets/AssetsService.test.ts
@@ -110,6 +110,7 @@ describe('AssetsService', () => {
       findAccountById: mockAccountsService.findById.bind(mockAccountsService),
       getActiveNetworks:
         mockConfigProvider.getActiveNetworks.bind(mockConfigProvider),
+      fetchMint: mockConnection.fetchMint.bind(mockConnection),
     });
 
     assetsService = new AssetsService({